Ran goyacc and gen new parser.go

+++ parser/parser.go
@@ -258,7 +258,7 @@ func operation(operator ast.OperatorType, left, right ast.Expression) (n ast.Exp
 }
 
 //line yacctab:1
-var yrExca = [...]int{
+var yrExca = [...]int16{
 	-1, 1,
 	1, -1,
 	-2, 15,
@@ -282,7 +282,7 @@ const yrPrivate = 57344
 
 const yrLast = 481
 
-var yrAct = [...]int{
+var yrAct = [...]int16{
 	52, 200, 199, 49, 68, 165, 166, 158, 83, 84,
 	85, 86, 87, 88, 89, 90, 249, 230, 228, 192,
 	250, 231, 229, 209, 104, 102, 103, 96, 97, 92,
@@ -334,7 +334,7 @@ var yrAct = [...]int{
 	217,
 }
 
-var yrPact = [...]int{
+var yrPact = [...]int16{
 	-1000, 208, -1000, -1000, 144, -1000, 291, 141, -1000, 266,
 	-1000, -1000, -1000, -1000, -1000, 7, 38, 265, 292, 247,
 	-1000, 278, 6, -1000, -1000, 5, 244, 276, 189, 244,
@@ -364,7 +364,7 @@ var yrPact = [...]int{
 	-1000, -1000,
 }
 
-var yrPgo = [...]int{
+var yrPgo = [...]int16{
 	0, 388, 387, 385, 380, 161, 374, 373, 372, 203,
 	369, 368, 367, 366, 365, 364, 363, 362, 355, 351,
 	3, 68, 0, 4, 342, 341, 137, 340, 339, 338,
@@ -372,7 +372,7 @@ var yrPgo = [...]int{
 	313, 312, 311,
 }
 
-var yrR1 = [...]int{
+var yrR1 = [...]int8{
 	0, 39, 39, 39, 39, 39, 1, 40, 41, 2,
 	7, 7, 8, 8, 19, 18, 18, 17, 17, 3,
 	3, 4, 4, 6, 6, 5, 5, 5, 5, 5,
@@ -391,7 +391,7 @@ var yrR1 = [...]int{
 	22, 22, 22,
 }
 
-var yrR2 = [...]int{
+var yrR2 = [...]int8{
 	0, 0, 2, 2, 3, 2, 2, 0, 0, 11,
 	0, 3, 0, 3, 3, 0, 2, 1, 1, 0,
 	2, 1, 2, 1, 2, 3, 3, 4, 3, 3,
@@ -410,7 +410,7 @@ var yrR2 = [...]int{
 	3, 3, 1,
 }
 
-var yrChk = [...]int{
+var yrChk = [...]int16{
 	-1000, -39, -2, -1, 52, 4, -18, 49, 21, 6,
 	-17, 7, 8, 21, 12, -40, -3, 77, 75, -4,
 	12, -7, 9, 12, -8, 10, 77, -41, 77, -6,
@@ -440,7 +440,7 @@ var yrChk = [...]int{
 	80, 80,
 }
 
-var yrDef = [...]int{
+var yrDef = [...]int16{
 	1, -2, 2, 3, 0, 5, 0, 0, 4, 0,
 	16, 17, 18, 6, 7, 19, 0, 0, 10, 20,
 	21, 12, 0, 22, 8, 0, 0, 0, 0, 11,
@@ -470,7 +470,7 @@ var yrDef = [...]int{
 	83, 49,
 }
 
-var yrTok1 = [...]int{
+var yrTok1 = [...]int8{
 	1, 3, 3, 3, 3, 3, 3, 3, 3, 3,
 	3, 3, 3, 3, 3, 3, 3, 3, 3, 3,
 	3, 3, 3, 3, 3, 3, 3, 3, 3, 3,
@@ -486,7 +486,7 @@ var yrTok1 = [...]int{
 	3, 3, 3, 75, 57, 76, 73,
 }
 
-var yrTok2 = [...]int{
+var yrTok2 = [...]int8{
 	2, 3, 4, 5, 6, 7, 8, 9, 10, 11,
 	12, 13, 14, 15, 16, 17, 18, 19, 20, 21,
 	22, 23, 24, 25, 26, 27, 28, 29, 30, 31,
@@ -496,7 +496,7 @@ var yrTok2 = [...]int{
 	65, 66, 67, 74,
 }
 
-var yrTok3 = [...]int{
+var yrTok3 = [...]int8{
 	0,
 }
 
@@ -578,9 +578,9 @@ func yrErrorMessage(state, lookAhead int) string {
 	expected := make([]int, 0, 4)
 
 	// Look for shiftable tokens.
-	base := yrPact[state]
+	base := int(yrPact[state])
 	for tok := TOKSTART; tok-1 < len(yrToknames); tok++ {
-		if n := base + tok; n >= 0 && n < yrLast && yrChk[yrAct[n]] == tok {
+		if n := base + tok; n >= 0 && n < yrLast && int(yrChk[int(yrAct[n])]) == tok {
 			if len(expected) == cap(expected) {
 				return res
 			}
@@ -590,13 +590,13 @@ func yrErrorMessage(state, lookAhead int) string {
 
 	if yrDef[state] == -2 {
 		i := 0
-		for yrExca[i] != -1 || yrExca[i+1] != state {
+		for yrExca[i] != -1 || int(yrExca[i+1]) != state {
 			i += 2
 		}
 
 		// Look for tokens that we accept or reduce.
 		for i += 2; yrExca[i] >= 0; i += 2 {
-			tok := yrExca[i]
+			tok := int(yrExca[i])
 			if tok < TOKSTART || yrExca[i+1] == 0 {
 				continue
 			}
@@ -627,30 +627,30 @@ func yrlex1(lex yrLexer, lval *yrSymType) (char, token int) {
 	token = 0
 	char = lex.Lex(lval)
 	if char <= 0 {
-		token = yrTok1[0]
+		token = int(yrTok1[0])
 		goto out
 	}
 	if char < len(yrTok1) {
-		token = yrTok1[char]
+		token = int(yrTok1[char])
 		goto out
 	}
 	if char >= yrPrivate {
 		if char < yrPrivate+len(yrTok2) {
-			token = yrTok2[char-yrPrivate]
+			token = int(yrTok2[char-yrPrivate])
 			goto out
 		}
 	}
 	for i := 0; i < len(yrTok3); i += 2 {
-		token = yrTok3[i+0]
+		token = int(yrTok3[i+0])
 		if token == char {
-			token = yrTok3[i+1]
+			token = int(yrTok3[i+1])
 			goto out
 		}
 	}
 
 out:
 	if token == 0 {
-		token = yrTok2[1] /* unknown char */
+		token = int(yrTok2[1]) /* unknown char */
 	}
 	if yrDebug >= 3 {
 		__yyfmt__.Printf("lex %s(%d)\n", yrTokname(token), uint(char))
@@ -705,7 +705,7 @@ yrstack:
 	yrS[yrp].yys = yrstate
 
 yrnewstate:
-	yrn = yrPact[yrstate]
+	yrn = int(yrPact[yrstate])
 	if yrn <= yrFlag {
 		goto yrdefault /* simple state */
 	}
@@ -716,8 +716,8 @@ yrnewstate:
 	if yrn < 0 || yrn >= yrLast {
 		goto yrdefault
 	}
-	yrn = yrAct[yrn]
-	if yrChk[yrn] == yrtoken { /* valid shift */
+	yrn = int(yrAct[yrn])
+	if int(yrChk[yrn]) == yrtoken { /* valid shift */
 		yrrcvr.char = -1
 		yrtoken = -1
 		yrVAL = yrrcvr.lval
@@ -730,7 +730,7 @@ yrnewstate:
 
 yrdefault:
 	/* default state action */
-	yrn = yrDef[yrstate]
+	yrn = int(yrDef[yrstate])
 	if yrn == -2 {
 		if yrrcvr.char < 0 {
 			yrrcvr.char, yrtoken = yrlex1(yrlex, &yrrcvr.lval)
@@ -739,18 +739,18 @@ yrdefault:
 		/* look through exception table */
 		xi := 0
 		for {
-			if yrExca[xi+0] == -1 && yrExca[xi+1] == yrstate {
+			if yrExca[xi+0] == -1 && int(yrExca[xi+1]) == yrstate {
 				break
 			}
 			xi += 2
 		}
 		for xi += 2; ; xi += 2 {
-			yrn = yrExca[xi+0]
+			yrn = int(yrExca[xi+0])
 			if yrn < 0 || yrn == yrtoken {
 				break
 			}
 		}
-		yrn = yrExca[xi+1]
+		yrn = int(yrExca[xi+1])
 		if yrn < 0 {
 			goto ret0
 		}
@@ -772,10 +772,10 @@ yrdefault:
 
 			/* find a state where "error" is a legal shift action */
 			for yrp >= 0 {
-				yrn = yrPact[yrS[yrp].yys] + yrErrCode
+				yrn = int(yrPact[yrS[yrp].yys]) + yrErrCode
 				if yrn >= 0 && yrn < yrLast {
-					yrstate = yrAct[yrn] /* simulate a shift of "error" */
-					if yrChk[yrstate] == yrErrCode {
+					yrstate = int(yrAct[yrn]) /* simulate a shift of "error" */
+					if int(yrChk[yrstate]) == yrErrCode {
 						goto yrstack
 					}
 				}
@@ -811,7 +811,7 @@ yrdefault:
 	yrpt := yrp
 	_ = yrpt // guard against "declared and not used"
 
-	yrp -= yrR2[yrn]
+	yrp -= int(yrR2[yrn])
 	// yrp is now the index of $0. Perform the default action. Iff the
 	// reduced production is ε, $1 is possibly out of range.
 	if yrp+1 >= len(yrS) {
@@ -822,16 +822,16 @@ yrdefault:
 	yrVAL = yrS[yrp+1]
 
 	/* consult goto table to find next state */
-	yrn = yrR1[yrn]
-	yrg := yrPgo[yrn]
+	yrn = int(yrR1[yrn])
+	yrg := int(yrPgo[yrn])
 	yrj := yrg + yrS[yrp].yys + 1
 
 	if yrj >= yrLast {
-		yrstate = yrAct[yrg]
+		yrstate = int(yrAct[yrg])
 	} else {
-		yrstate = yrAct[yrj]
-		if yrChk[yrstate] != -yrn {
-			yrstate = yrAct[yrg]
+		yrstate = int(yrAct[yrj])
+		if int(yrChk[yrstate]) != -yrn {
+			yrstate = int(yrAct[yrg])
 		}
 	}
 	// dummy call; replaced with literal code
